requirements.txt
langchain_google_genai
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
google-cloud-aiplatform>=1.38
selenium
langchain_community
qdrant-client
unstructured
fastapi
